added reset password

+++ matcha-express_new/src/service/authService.ts
@@ -4,6 +4,8 @@ import { emailVerificationService } from "./emailVerificationService";
 import mapDbError from "../utils/mapDbError";
 import { UnauthorizedError } from "../utils/errors";
 import bcrypt from "bcryptjs";
+import crypto from "crypto";
+import nodemailer from "nodemailer";
 
 export const authService = {
 	async register(user: IUser): Promise<Omit<IUser, "password">> {
@@ -13,9 +15,9 @@ export const authService = {
 			const token = await emailVerificationService.generateEmailVerificationToken();
 			if (!newUser || !newUser.id) throw new UnauthorizedError("User creation failed");
 			await emailVerificationService.createEmailVerification(newUser.id, token);
-			await emailVerificationService.sendVerificationEmail(user.email, token);
+			await emailVerificationService.sendVerificationEmail(user.email, token, {text: " to verify your email.", subject: "Email Verification", url: "verify-email"});
 
-			const { password: _pw, ...safeUser } = newUser
+			const { password: _pw, ...safeUser } = newUser;
 
 			return safeUser;
 		} catch (error: any) {
@@ -44,16 +46,54 @@ export const authService = {
 		}
 	},
 
-	async verifyEmail(token: string): Promise<{ message: string, username: string }> {
-		const emailVerification = await emailVerificationService.findByToken(token);
-		if (!emailVerification) throw new UnauthorizedError("Invalid or expired token");
-		if (!emailVerification.id) throw new UnauthorizedError("Invalid token");
-		// if (emailVerification.expires_at && emailVerification.expires_at < new Date()) throw new UnauthorizedError("Token expired");
-		const userId = emailVerification.user_id;
-		const user = await userService.getUserById(userId)
-		await userService.updateUser(userId, { email_verified: true });
-		await emailVerificationService.deleteEmailVerification(emailVerification.id);
-		return { message: "Email verified successfully", username: user?.username! };
+	async generateResetToken(): Promise<string> {
+		return crypto.randomBytes(32).toString("hex");
 	},
 
+	async verifyEmail(token: string): Promise<{ message: string; username: string }> {
+		try {
+			const emailVerification = await emailVerificationService.findByToken(token);
+			if (!emailVerification) throw new UnauthorizedError("Invalid or expired token");
+			if (!emailVerification.id) throw new UnauthorizedError("Invalid token");
+			if (emailVerification.expires_at && emailVerification.expires_at < new Date()) throw new UnauthorizedError("Token expired");
+			const userId = emailVerification.user_id;
+			const user = await userService.getUserById(userId);
+			await userService.updateUser(userId, { email_verified: true });
+			await emailVerificationService.deleteEmailVerification(emailVerification.id);
+			return { message: "Email verified successfully", username: user?.username! };
+		} catch (error: any) {
+			throw mapDbError.auth(error);
+		}
+	},
+
+	async resetPassword(email: string): Promise<{ message: string }> {
+		try {
+			const user = await userService.getUserByEmail(email);
+			if (!user) throw new UnauthorizedError("User not found");
+			const token = await emailVerificationService.generateEmailVerificationToken();
+			await emailVerificationService.createEmailVerification(user.id!, token);
+			await emailVerificationService.sendVerificationEmail(email, token, {text: " to reset your password.", subject: "Password Reset", url: "reset-password"});
+			return { message: "Password reset email sent" };
+		} catch (error: any) {
+			throw mapDbError.auth(error);
+		}
+	},
+
+	async verifyResetPassword(token: string, password: string): Promise<{ message: string }> {
+		try {
+			const emailVerification = await emailVerificationService.findByToken(token);
+			if (!emailVerification) throw new UnauthorizedError("Invalid or expired token");
+			if (!emailVerification.id) throw new UnauthorizedError("Invalid token");
+			if (emailVerification.expires_at && emailVerification.expires_at < new Date()) throw new UnauthorizedError("Token expired");
+			const userId = emailVerification.user_id;
+			const user = await userService.getUserById(userId);
+			if (!user) throw new UnauthorizedError("User not found");
+			user.password = await bcrypt.hash(password, 10);
+			await userService.updateUser(userId, { password: user.password });
+			await emailVerificationService.deleteEmailVerification(emailVerification.id);
+			return { message: "Password reset successfully" };
+		} catch (error: any) {
+			throw mapDbError.auth(error);
+		}
+	},
 };
